Social media active user actions like other apps
This application offer sign-up , sign-in and post some photograph. when you create an account you can like photo whose belongs to other users , and you can save their post as the same with liking thing, you can comment it and if you want to delete it , you can. following and unfollowing process works, briefly their data who is sharing with the application store Firebase in Cloud firestore and other authencation process will store as you typed to app in firebase.